Update git configuration
- Remove email address from git config
- Disable diff noprefix setting
- Clean up global gitignore
- Add git template directory for talisman hooks
- Add git.zsh with GIT_PAGER and GIT_EDITOR settings

+++ install.sh.d/git.sh
@@ -1,26 +1,34 @@
 #!/bin/bash -e
+# DEPENDS: cargo stow wget
 
 if [ "$(uname -s)" == "Darwin" ]; then
     if ! brew ls --versions git >/dev/null ; then
         brew install git
     fi
-    if ! brew ls --versions zaquestion/tap/lab >/dev/null ; then
-        brew install zaquestion/tap/lab
-    fi
-    if ! [ -f $HOME/.cargo/bin/git-absorb ]; then
-        sudo apt install --assume-yes cargo
-        cargo install git-absorb
+elif [ "$(uname -s)" == "Linux" ]; then
+    if ! dpkg -s git >/dev/null ; then
+        sudo apt-get update
+        sudo apt-get install --yes git
     fi
 fi
 
+# Extension git-absorb
+if ! [ -f $HOME/.cargo/bin/git-absorb ]; then
+    [ -f $HOME/.cargo/bin/cargo install git-absorb
+fi
+
+# Extension git-diff-image
 if [ "$(uname -s)" == "Linux" ]; then
-    sudo apt-get install --yes exiftool imagemagick xdg-utils
+    if ! dpkg -s exiftool imagemagick xdg-utils >/dev/null ; then
+        sudo apt-get update
+        sudo apt-get install --yes exiftool imagemagick xdg-utils
+    fi
 fi
-
-stow git
-
-# Install git-diff-image
-SHA1=12e717f5e9b084e117c354db5204b949b3232a95
+SHA1=f12098b2b9b9f56f205f8e9ca8435796a0fdc1fc
 DEST=~/.local/bin
+mkdir -p ${DEST}
 wget https://raw.githubusercontent.com/ewanmellor/git-diff-image/${SHA1}/diff-image -O ${DEST}/diff-image
 wget https://raw.githubusercontent.com/ewanmellor/git-diff-image/${SHA1}/git_diff_image -O ${DEST}/git_diff_image
+
+# Install config files
+stow git
